rename scatter and browser folders
