+++ server/templates/browser/node.html
@@ -16,7 +16,7 @@
 
    {% extends 'base.html' %}
 
-   {% set main_id = 'dc-browser2' %}
+   {% set main_id = 'dc-browser' %}
    {% set title = node_name + ' - Graph Browser' %}
    {% set subpage_title = 'Graph Browser' %}
    {% set subpage_url = url_for('browser.browser_main') %}
