+++ static/js/browser/arc_table_row.tsx
