+++ static/js/browser/constants.ts
