+++ static/js/browser/image_section.tsx
