+++ static/js/browser/in_arc_section.tsx
