+++ static/js/browser/observation_chart.tsx
