+++ static/js/browser/observation_chart_section.tsx
