+++ static/js/browser/out_arc_section.tsx
