+++ static/js/browser/statvar_charts.tsx
