+++ static/js/browser/statvar_checkbox.tsx
