+++ static/js/browser/statvar_group_node.tsx
