+++ static/js/browser/statvar_group_section.tsx
