+++ static/js/browser/statvar_hierarchy.tsx
