+++ static/js/browser/statvar_hierarchy_node_header.tsx
