+++ static/js/browser/statvar_hierarchy_search.tsx
