+++ static/js/browser/statvar_section.tsx
