+++ static/js/browser/util.ts
