+++ static/js/browser/weather_chart_section.tsx
