+++ static/js/tools/scatter/app.tsx
