+++ static/js/tools/scatter/chart.tsx
