+++ static/js/tools/scatter/chart_loader.tsx
