+++ static/js/tools/scatter/context.ts
