+++ static/js/tools/scatter/info.tsx
