+++ static/js/tools/scatter/place_options.tsx
