+++ static/js/tools/scatter/plot_options.tsx
