+++ static/js/tools/scatter/scatter.ts
