+++ static/js/tools/scatter/spinner.tsx
