+++ static/js/tools/scatter/statvar.tsx
