+++ static/js/tools/scatter/util.ts
